requirements.txt
# Core Machine Learning
autogluon.tabular>=1.0.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0

# Deep Learning and Transformers
torch>=2.0.0
transformers>=4.30.0

# LLM API
openai>=1.0.0

# Feature Engineering
holidays>=0.30

# Utilities
psutil>=5.9.0
python-dotenv>=1.0.0

# Optional: Visualization (uncomment if needed)
# matplotlib>=3.7.0
# seaborn>=0.12.0

# Development Tools (uncomment for development)
# jupyter>=1.0.0
# ipython>=8.0.0
# pytest>=7.3.0

